kairos/requirements.txt
croniter>=2.0.0
PyYAML>=6.0
httpx>=0.27.0
python-frontmatter>=1.0.0
pytest>=8.0.0
